refactor(telegram): extract s1 clusters into mixin modules (epic #78791)
Verbatim extraction of 53 methods across rich_mixin, topics_mixin,
init_guards, network_mixin, text_format, config_mixin, authz_mixin
(4 authz methods skipped — covered by #75742). Wave-2 VERIFIED by
w2a/w2b, wave-3 VERIFIED by w3a/w3b. 101 regression tests, 509/509
baseline-identical on the full telegram suite.
