+++ plugins/platforms/telegram/adapter.py
@@ -24,6 +24,8 @@ from typing import Dict, List, Optional, Set, Any
 
 logger = logging.getLogger(__name__)
 
+from plugins.platforms.telegram.authz_mixin import _scoped_gate_env
+
 
 def _redact_telegram_error_text(error: object) -> str:
     """Redact secrets from Telegram transport errors before logging or returning them."""
@@ -37,201 +39,6 @@ def _redact_telegram_error_text(error: object) -> str:
     except Exception:
         return "<telegram error redacted>"
 
-
-def _scoped_gate_env(name: str, default: str = "") -> str:
-    """Read a TELEGRAM_*/GATEWAY_* authorization gate env var per-profile.
-
-    Under gateway.multiplex_profiles the process env is first-writer-wins
-    (the YAML→env bridge in ``_apply_yaml_config``), so a raw ``os.getenv``
-    can return ANOTHER profile's allowlist (issue #72348, Telegram mirror).
-    Reads the active profile's secret scope when installed; falls back to
-    ``os.getenv`` outside multiplex — identical single-profile behavior.
-    """
-    try:
-        from gateway.authz_mixin import _platform_gate_env
-
-        return _platform_gate_env(name, default)
-    except Exception:
-        return (os.getenv(name) or default).strip()
-
-
-def _consume_abandoned_task(task: asyncio.Task) -> None:
-    """Observe a detached task's terminal exception to avoid noisy loop logs."""
-    try:
-        task.exception()
-    except asyncio.CancelledError:
-        pass
-    except Exception:
-        logger.debug("Abandoned Telegram init task failed after timeout", exc_info=True)
-
-
-# Grace period after the wall-clock deadline fires: if the event loop still
-# hasn't processed the expiry callback by then, the loop thread itself is
-# blocked in a synchronous call — the exact state in which every asyncio-based
-# timeout (including this helper's own expiry hand-off) goes silent, so the
-# gateway hangs at "attempt 1/8" with no further output (#63309).
-_LOOP_BLOCKED_DUMP_GRACE = 5.0
-
-
-def _dump_loop_blocked_diagnostics(timeout: float, grace: float) -> None:
-    """Emit diagnostics from the deadline timer thread when the loop is stuck.
-
-    Runs OFF the event loop, so it works precisely when the loop cannot. The
-    faulthandler dump names the frame the loop thread is blocked in — the one
-    piece of information #63309-class hangs otherwise never surface.
-    """
-    logger.warning(
-        "[Telegram] init deadline (%.0fs) expired but the event loop has not "
-        "processed the expiry after a further %.0fs — the loop thread appears "
-        "BLOCKED in a synchronous call, which is why no timeout fires (#63309). "
-        "Dumping all thread stacks to stderr to identify the blocking frame.",
-        timeout,
-        grace,
-    )
-    try:
-        faulthandler.dump_traceback(all_threads=True)
-    except Exception:
-        logger.debug("faulthandler traceback dump failed", exc_info=True)
-
-
-async def _await_with_thread_deadline(awaitable, timeout: float, *, on_abandon=None):
-    """Await with a wall-clock deadline that does not depend on loop timers.
-
-    ``asyncio.wait_for`` schedules its timeout on the event loop and then waits
-    for cancellation to propagate.  PTB/httpcore initialization can sit inside
-    cancellation-shielded anyio scopes, so a timed-out initialize() may never
-    hand control back to the retry ladder under some supervisors.  This helper
-    lets a daemon ``threading.Timer`` wake the loop and, on timeout, abandons
-    the shielded task instead of awaiting cancellation completion.
-
-    ``on_abandon`` (optional) is a zero-arg callable returning an awaitable that
-    is scheduled as a detached best-effort cleanup when the task is abandoned on
-    timeout.  The abandoned initialize() may leave a half-built httpx client /
-    connection pool open (it never completed and we do not await its
-    cancellation), so the caller uses this to shut that state down and avoid
-    leaking a pool per retry attempt.  Cleanup runs detached and its own errors
-    are swallowed, so it can never re-block the retry ladder.
-    """
-    task = asyncio.ensure_future(awaitable)
-    loop = asyncio.get_running_loop()
-    deadline = loop.create_future()
-    # Set the moment the loop actually runs the expiry callback (or the helper
-    # exits normally). threading.Event so the watchdog thread can read it
-    # without touching asyncio state from off-loop.
-    loop_processed_expiry = threading.Event()
-
-    def _mark_expired() -> None:
-        loop_processed_expiry.set()
-        if not deadline.done():
-            deadline.set_result(None)
-
-    def _expire_from_thread() -> None:
-        loop.call_soon_threadsafe(_mark_expired)
-
-    def _watchdog_check() -> None:
-        # The deadline fired _LOOP_BLOCKED_DUMP_GRACE ago but the loop never
-        # ran _mark_expired: the loop thread is stuck in a synchronous call.
-        # Diagnose from this thread — the loop can't.
-        if not loop_processed_expiry.is_set():
-            _dump_loop_blocked_diagnostics(timeout, _LOOP_BLOCKED_DUMP_GRACE)
-
-    timer = threading.Timer(max(timeout, 0.0), _expire_from_thread)
-    timer.daemon = True
-    timer.start()
-    watchdog = threading.Timer(
-        max(timeout, 0.0) + _LOOP_BLOCKED_DUMP_GRACE, _watchdog_check
-    )
-    watchdog.daemon = True
-    watchdog.start()
-    try:
-        done, _ = await asyncio.wait(
-            {task, deadline},
-            return_when=asyncio.FIRST_COMPLETED,
-        )
-        if task in done:
-            if not deadline.done():
-                deadline.cancel()
-            return await task
-
-        task.cancel()
-        task.add_done_callback(_consume_abandoned_task)
-        if on_abandon is not None:
-            # Detached best-effort cleanup: close the half-built app's httpx
-            # client/pool so an abandoned attempt can't leak sockets across the
-            # retry ladder. Detached + exception-observed so it never re-blocks
-            # or re-hangs the ladder we are trying to advance.
-            cleanup = asyncio.ensure_future(_run_abandon_cleanup(on_abandon))
-            cleanup.add_done_callback(_consume_abandoned_task)
-        raise asyncio.TimeoutError()
-    finally:
-        timer.cancel()
-        watchdog.cancel()
-        # cancel() cannot stop a Timer whose callback is already running;
-        # setting the event closes that race so a completed await can never
-        # be misreported as a blocked loop.
-        loop_processed_expiry.set()
-
-
-async def _first_completed(*futures: "asyncio.Future") -> None:
-    """Return when the first of ``futures`` completes.
-
-    Used by the strict cold-start readiness gate to wait on "progress OR
-    polling error", whichever fires first (#67498). Does not cancel the
-    losers — the caller owns their lifecycle.
-    """
-    await asyncio.wait(set(futures), return_when=asyncio.FIRST_COMPLETED)
-
-
-async def _run_abandon_cleanup(on_abandon) -> None:
-    """Run the abandonment cleanup coroutine, swallowing any failure.
-
-    Wrapped so a cleanup that itself hangs or raises cannot surface as an
-    unhandled task error or block anything — it is fully fire-and-forget.
-    """
-    try:
-        result = on_abandon()
-        if asyncio.iscoroutine(result) or asyncio.isfuture(result):
-            await result
-    except Exception:
-        logger.debug("Abandoned Telegram init cleanup failed", exc_info=True)
-
-
-async def _shutdown_abandoned_app(app) -> None:
-    """Release a half-built PTB app's httpx transports after init was abandoned.
-
-    ``Application.shutdown()`` / ``Bot.shutdown()`` are gated on the app's
-    ``_initialized`` / ``_requests_initialized`` flags, which a wedged
-    ``initialize()`` (the case this whole path exists for) may never have set —
-    so calling only ``app.shutdown()`` no-ops and leaks the connection pool it
-    was meant to close.  ``HTTPXRequest`` builds its ``httpx.AsyncClient``
-    eagerly in its constructor and its ``shutdown()`` gates only on
-    ``client.is_closed``, so closing the request transports directly releases
-    the pool regardless of PTB init state.  We try the clean path first, then
-    fall back to the transports.  All best-effort and swallowed.
-    """
-    if app is None:
-        return
-    try:
-        await app.shutdown()
-    except Exception:
-        logger.debug("Abandoned Telegram app.shutdown() failed", exc_info=True)
-    # Directly close the underlying request transports (bypasses PTB's
-    # init-gated shutdown so the eagerly-built httpx pool is released even when
-    # the abandoned initialize() never flipped _initialized).
-    bot = getattr(app, "bot", None)
-    requests = getattr(bot, "_request", None) if bot is not None else None
-    if not requests:
-        return
-    for request in requests:
-        shutdown = getattr(request, "shutdown", None)
-        if shutdown is None:
-            continue
-        try:
-            result = shutdown()
-            if asyncio.iscoroutine(result) or asyncio.isfuture(result):
-                await result
-        except Exception:
-            logger.debug("Abandoned Telegram request shutdown failed", exc_info=True)
 
 try:
     from telegram import Update, Bot, Message, InlineKeyboardButton, InlineKeyboardMarkup
@@ -305,6 +112,25 @@ from plugins.platforms.telegram.telegram_network import (
     discover_fallback_ips,
     parse_fallback_ip_env,
 )
+from plugins.platforms.telegram.authz_mixin import TelegramAuthorizationMixin
+from plugins.platforms.telegram.config_mixin import ConfigMixin
+from plugins.platforms.telegram.init_guards import (
+    _LOOP_BLOCKED_DUMP_GRACE,
+    _await_with_thread_deadline,
+    _consume_abandoned_task,
+    _dump_loop_blocked_diagnostics,
+    _first_completed,
+    _run_abandon_cleanup,
+    _shutdown_abandoned_app,
+)
+from plugins.platforms.telegram.network_mixin import NetworkMixin
+from plugins.platforms.telegram.rich_mixin import RichMixin
+from plugins.platforms.telegram.text_format import (
+    _escape_mdv2,
+    _separate_chunk_indicator_from_fence,
+    _strip_mdv2,
+)
+from plugins.platforms.telegram.topics_mixin import TopicsMixin
 from utils import atomic_replace, env_float, env_int
 
 _TELEGRAM_IMAGE_EXTENSIONS = {".png", ".jpg", ".jpeg", ".webp", ".gif"}
@@ -447,56 +273,6 @@ def check_telegram_requirements() -> bool:
     return True
 
 
-# Matches every character that MarkdownV2 requires to be backslash-escaped
-# when it appears outside a code span or fenced code block.
-_MDV2_ESCAPE_RE = re.compile(r'([_*\[\]()~`>#\+\-=|{}.!\\])')
-
-
-def _escape_mdv2(text: str) -> str:
-    """Escape Telegram MarkdownV2 special characters with a preceding backslash."""
-    return _MDV2_ESCAPE_RE.sub(r'\\\1', text)
-
-
-def _strip_mdv2(text: str) -> str:
-    """Strip MarkdownV2 escape backslashes to produce clean plain text.
-
-    Also removes MarkdownV2 formatting markers so the fallback
-    doesn't show stray syntax characters from format_message conversion.
-    """
-    # Remove escape backslashes before special characters
-    cleaned = re.sub(r'\\([_*\[\]()~`>#\+\-=|{}.!\\])', r'\1', text)
-    # Remove standard markdown bold (**text** → text) BEFORE MarkdownV2 bold
-    cleaned = re.sub(r'\*\*([^*]+)\*\*', r'\1', cleaned)
-    # Remove MarkdownV2 bold markers that format_message converted from **bold**
-    cleaned = re.sub(r'\*([^*]+)\*', r'\1', cleaned)
-    # Remove MarkdownV2 italic markers that format_message converted from *italic*
-    # Use word boundary (\b) to avoid breaking snake_case like my_variable_name
-    cleaned = re.sub(r'(?<!\w)_([^_]+)_(?!\w)', r'\1', cleaned)
-    # Remove MarkdownV2 strikethrough markers (~text~ → text)
-    cleaned = re.sub(r'~([^~]+)~', r'\1', cleaned)
-    # Remove MarkdownV2 spoiler markers (||text|| → text)
-    cleaned = re.sub(r'\|\|([^|]+)\|\|', r'\1', cleaned)
-    return cleaned
-
-
-_CHUNK_INDICATOR_ON_FENCE_RE = re.compile(
-    r'(?m)^``` (?P<indicator>(?:\\)?\(\d+/\d+(?:\\)?\))$'
-)
-
-
-def _separate_chunk_indicator_from_fence(text: str) -> str:
-    """Move ``(N/M)`` chunk markers off Telegram code-fence lines.
-
-    ``truncate_message()`` appends chunk indicators to the end of a chunk. When
-    the chunk had to close an in-progress fenced code block, that creates a
-    line like ````` \\(1/2\\)`` after MarkdownV2 escaping. Telegram does not
-    treat that as a clean closing fence, so it can reject MarkdownV2 and fall
-    back to plain text. Put the indicator on its own line immediately after the
-    closing fence.
-    """
-    return _CHUNK_INDICATOR_ON_FENCE_RE.sub(r'```\n\g<indicator>', text)
-
-
 # ---------------------------------------------------------------------------
 # Markdown table → Telegram-friendly row groups
 # ---------------------------------------------------------------------------
@@ -510,55 +286,6 @@ from gateway.platforms.helpers import (
     compile_mention_patterns,
     convert_table_to_bullets as _wrap_markdown_tables,
 )
-
-
-# ---------------------------------------------------------------------------
-# Rich-message newline normalization
-# ---------------------------------------------------------------------------
-
-# Matches a protected region whose internal newlines must stay bare in the
-# rich-message path: a fenced code block (```...```) OR a GFM pipe-table block
-# (a header row, a delimiter row of dashes/pipes, then any pipe data rows).
-# Telegram renders both natively, so injecting Markdown hard breaks inside them
-# would corrupt the code block / table.
-_RICH_PROTECTED_REGION_RE = re.compile(
-    r'(?:```[^\n]*\n[\s\S]*?```)'                       # fenced code block
-    r'|(?:^[^\n]*\|[^\n]*\n'                            # table header row (has a pipe)
-    r'[ \t]*\|?[ \t]*:?-+:?[ \t]*(?:\|[ \t]*:?-+:?[ \t]*)+\|?[ \t]*'  # delimiter
-    r'(?:\n[^\n]*\|[^\n]*)*)',                          # data rows (newline-led, trailing \n left for prose)
-    re.MULTILINE,
-)
-
-
-def _rich_normalize_linebreaks(text: str) -> str:
-    """Convert single ``\\n`` to Markdown hard breaks for the rich-message path.
-
-    Standard Markdown treats a lone ``\\n`` as whitespace (soft break), so
-    Bot API 10.1 ``sendRichMessage`` collapses multi-line content — e.g.
-    slash-command lists joined with ``"\\n".join(lines)`` — into a single
-    paragraph.  Adding two trailing spaces before each single newline
-    forces a hard line break (``<br>``) in the rendered output.
-
-    Paragraph breaks (``\\n\\n``), fenced code blocks, and GFM pipe-table
-    blocks are left untouched: tables render natively in the rich path and a
-    hard break injected into a row separator would corrupt the table.
-    """
-    if not text or '\n' not in text:
-        return text
-
-    out: list[str] = []
-    # Split off protected regions (fenced code OR table blocks) and only inject
-    # hard breaks in the prose between them. Boundary newlines are handled by
-    # the original single-\n regex, which sees each prose run as a whole string.
-    pos = 0
-    for m in _RICH_PROTECTED_REGION_RE.finditer(text):
-        prose = text[pos:m.start()]
-        out.append(re.sub(r'(?<!\n)\n(?!\n)', '  \n', prose))
-        out.append(m.group(0))  # protected region kept verbatim
-        pos = m.end()
-    tail = text[pos:]
-    out.append(re.sub(r'(?<!\n)\n(?!\n)', '  \n', tail))
-    return ''.join(out)
 
 
 # Watchdog bound for `await updater.stop()`. When the underlying TCP socket is
@@ -614,7 +341,14 @@ class _PollingLifecycleAbort(RuntimeError):
     """Internal control flow for polling startup fenced by teardown."""
 
 
-class TelegramAdapter(BasePlatformAdapter):
+class TelegramAdapter(
+    TelegramAuthorizationMixin,
+    ConfigMixin,
+    NetworkMixin,
+    TopicsMixin,
+    RichMixin,
+    BasePlatformAdapter,
+):
     """
     Telegram bot adapter.
 
@@ -666,34 +400,6 @@ class TelegramAdapter(BasePlatformAdapter):
     _TEXT_BATCH_FAST_DELAY_S = 0.18
     _TEXT_BATCH_SHORT_LEN = 1024
     _TEXT_BATCH_SHORT_DELAY_S = 0.24
-
-    @staticmethod
-    def _env_float_clamped(
-        name: str,
-        default: float,
-        *,
-        min_value: Optional[float] = None,
-        max_value: Optional[float] = None,
-    ) -> float:
-        """Read a float env var, reject non-finite values, and clamp to bounds.
-
-        Guarantees the returned value is a finite number usable directly in
-        ``asyncio.sleep()`` and similar APIs that reject NaN / Inf.
-        """
-        import math
-
-        raw = os.getenv(name)
-        try:
-            value = float(raw) if raw is not None else float(default)
-        except (TypeError, ValueError):
-            value = float(default)
-        if not math.isfinite(value):
-            value = float(default)
-        if min_value is not None:
-            value = max(value, min_value)
-        if max_value is not None:
-            value = min(value, max_value)
-        return value
 
     @property
     def message_len_fn(self):
@@ -1047,40 +753,6 @@ class TelegramAdapter(BasePlatformAdapter):
         )
         return any(_scoped_gate_env(key).strip() for key in keys)
 
-    def _should_pass_unauthorized_dm_for_pairing(self, source) -> bool:
-        """Return True when an unauthorized DM must still reach gateway pairing.
-
-        Early auth (#40863) rejects before event construction. That is correct
-        when unauthorized DMs are ignored, but it must not short-circuit the
-        gateway pairing handshake when ``unauthorized_dm_behavior`` resolves
-        to ``pair`` — including the case where an allowlist is set and the
-        operator explicitly opted back into pairing via a platform override
-        (resolution rule 1 in ``_get_unauthorized_dm_behavior``).
-        """
-        if source.chat_type != "dm":
-            return False
-
-        runner = getattr(getattr(self, "_message_handler", None), "__self__", None)
-        behavior_fn = getattr(runner, "_get_unauthorized_dm_behavior", None)
-        if callable(behavior_fn):
-            try:
-                return (
-                    behavior_fn(
-                        Platform.TELEGRAM,
-                        profile=getattr(source, "profile", None),
-                    )
-                    == "pair"
-                )
-            except Exception:
-                logger.debug(
-                    "[Telegram] Failed to resolve unauthorized DM behavior; "
-                    "falling back to adapter-local override",
-                    exc_info=True,
-                )
-
-        extra = getattr(getattr(self, "config", None), "extra", None) or {}
-        return str(extra.get("unauthorized_dm_behavior", "")).strip().lower() == "pair"
-
     def _is_user_authorized_from_message(self, message: Message) -> bool:
         """Check if the sender of a Telegram message is authorized.
 
@@ -1166,870 +838,12 @@ class TelegramAdapter(BasePlatformAdapter):
         # Unauthorized DM that the gateway would pair: forward so pairing can run.
         return self._should_pass_unauthorized_dm_for_pairing(source)
 
-    @classmethod
-    def _metadata_thread_id(cls, metadata: Optional[Dict[str, Any]]) -> Optional[str]:
-        if not metadata:
-            return None
-        thread_id = metadata.get("thread_id") or metadata.get("message_thread_id")
-        return str(thread_id) if thread_id is not None else None
-
-    @classmethod
-    def _metadata_direct_messages_topic_id(cls, metadata: Optional[Dict[str, Any]]) -> Optional[str]:
-        if not metadata:
-            return None
-        topic_id = metadata.get("direct_messages_topic_id") or metadata.get("telegram_direct_messages_topic_id")
-        return str(topic_id) if topic_id is not None else None
-
-    @classmethod
-    def _metadata_reply_to_message_id(cls, metadata: Optional[Dict[str, Any]]) -> Optional[int]:
-        if not metadata:
-            return None
-        reply_to = metadata.get("telegram_reply_to_message_id")
-        return int(reply_to) if reply_to is not None else None
-
-    @classmethod
-    def _is_private_dm_topic_send(
-        cls,
-        chat_id: str,
-        thread_id: Optional[str],
-        metadata: Optional[Dict[str, Any]],
-    ) -> bool:
-        if cls._metadata_direct_messages_topic_id(metadata) is not None:
-            return bool(
-                metadata
-                and metadata.get("telegram_dm_topic_reply_fallback")
-                and cls._metadata_reply_to_message_id(metadata) is not None
-            )
-        if metadata and metadata.get("telegram_dm_topic_created_for_send"):
-            return False
-        return bool(
-            thread_id
-            and metadata
-            and metadata.get("telegram_dm_topic_reply_fallback")
-        )
-
-    @staticmethod
-    def _dm_topic_missing_anchor_error() -> str:
-        return "Telegram DM topic delivery requires a reply anchor; refusing to send outside the requested topic"
-
-    @classmethod
-    def _reply_to_message_id_for_send(
-        cls,
-        reply_to: Optional[str],
-        metadata: Optional[Dict[str, Any]] = None,
-        reply_to_mode: Optional[str] = None,
-    ) -> Optional[int]:
-        if reply_to:
-            return int(reply_to)
-        if metadata and metadata.get("telegram_dm_topic_reply_fallback"):
-            if reply_to_mode == "off":
-                return None
-            return cls._metadata_reply_to_message_id(metadata)
-        return None
-
-    @classmethod
-    def _thread_kwargs_for_send(
-        cls,
-        chat_id: str,
-        thread_id: Optional[str],
-        metadata: Optional[Dict[str, Any]] = None,
-        reply_to_message_id: Optional[int] = None,
-        reply_to_mode: Optional[str] = None,
-    ) -> Dict[str, Any]:
-        """Return Telegram send kwargs for forum and direct-message topic routing.
-
-        Supergroup/forum topics use ``message_thread_id``. True Bot API Direct
-        Messages topics can opt in with explicit ``direct_messages_topic_id``
-        metadata. Hermes-created private-chat topic lanes are marked with
-        ``telegram_dm_topic_reply_fallback``. Live replies send the private
-        topic thread id together with a reply anchor; synthetic/resumed sends
-        without an anchor use ``direct_messages_topic_id`` when metadata has it.
-        ``message_thread_id`` alone can render outside the visible lane.
-
-        When ``reply_to_mode`` is ``"off"``, the reply anchor is suppressed for
-        DM topic fallback sends while preserving the ``message_thread_id`` so
-        the message still lands in the correct topic.
-        """
-        if metadata and metadata.get("telegram_dm_topic_reply_fallback"):
-            if reply_to_mode == "off":
-                return {"message_thread_id": cls._message_thread_id_for_send(thread_id)}
-            if reply_to_message_id is None:
-                reply_to_message_id = cls._metadata_reply_to_message_id(metadata)
-            if reply_to_message_id is None:
-                direct_topic_id = cls._metadata_direct_messages_topic_id(metadata)
-                if direct_topic_id is not None:
-                    return {
-                        "message_thread_id": None,
-                        "direct_messages_topic_id": int(direct_topic_id),
-                    }
-                return {}
-            return {"message_thread_id": cls._message_thread_id_for_send(thread_id)}
-        direct_topic_id = cls._metadata_direct_messages_topic_id(metadata)
-        if direct_topic_id is not None:
-            return {
-                "message_thread_id": None,
-                "direct_messages_topic_id": int(direct_topic_id),
-            }
-        return {"message_thread_id": cls._message_thread_id_for_send(thread_id)}
-
-    @classmethod
-    def _message_thread_id_for_send(cls, thread_id: Optional[str]) -> Optional[int]:
-        if not thread_id or str(thread_id) == cls._GENERAL_TOPIC_THREAD_ID:
-            return None
-        return int(thread_id)
-
-    @classmethod
-    def _message_thread_id_for_typing(cls, thread_id: Optional[str]) -> Optional[int]:
-        # Asymmetric with _message_thread_id_for_send on purpose. Telegram's
-        # sendMessage and sendChatAction treat thread id "1" (the forum General
-        # topic) differently: sends reject message_thread_id=1 and must omit it,
-        # but sendChatAction needs message_thread_id=1 to place the typing
-        # bubble in the General topic (omitting it hides the bubble entirely
-        # from the client's view of that topic). Preserve the real id here —
-        # sends still map "1" → None via _message_thread_id_for_send.
-        if not thread_id:
-            return None
-        return int(thread_id)
-
-    @staticmethod
-    def _is_thread_not_found_error(error: Exception) -> bool:
-        return "thread not found" in str(error).lower()
-
-    def _prune_stale_dm_topic_binding(
-        self, chat_id: Any, thread_id: Any,
-    ) -> None:
-        """Drop the stale ``telegram_dm_topic_bindings`` row for a
-        topic Telegram has confirmed deleted.
-
-        Without this prune the recovery logic in
-        ``gateway.run._recover_telegram_topic_thread_id`` keeps
-        steering future inbound messages to the dead thread (the
-        bug behind #31501 — tool progress, approvals, replies all
-        end up in the wrong place even though the user has moved
-        on to a fresh topic).  Best-effort: we never raise from a
-        send-fallback path — a failed cleanup must not turn into a
-        failed user-facing send.
-        """
-        if chat_id is None or thread_id is None:
-            return
-        store = getattr(self, "_session_store", None)
-        if store is None:
-            return
-        db = getattr(store, "_db", None)
-        if db is None or not hasattr(db, "delete_telegram_topic_binding"):
-            return
-        try:
-            removed = db.delete_telegram_topic_binding(
-                chat_id=str(chat_id), thread_id=str(thread_id),
-            )
-        except Exception:
-            logger.debug(
-                "[%s] delete_telegram_topic_binding failed for "
-                "chat=%s thread=%s — skipping prune",
-                self.name, chat_id, thread_id, exc_info=True,
-            )
-            return
-        if removed:
-            logger.info(
-                "[%s] Pruned stale Telegram DM topic binding "
-                "chat=%s thread=%s (Bot API: thread not found)",
-                self.name, chat_id, thread_id,
-            )
-
-    @staticmethod
-    def _is_bad_request_error(error: Exception) -> bool:
-        name = error.__class__.__name__.lower()
-        if name == "badrequest" or name.endswith("badrequest"):
-            return True
-        try:
-            from telegram.error import BadRequest
-            return isinstance(error, BadRequest)
-        except ImportError:
-            return False
-
-    @classmethod
-    def _should_retry_without_dm_topic_reply_anchor(
-        cls,
-        error: Exception,
-        metadata: Optional[Dict[str, Any]],
-        reply_to_message_id: Optional[int],
-    ) -> bool:
-        """True when a DM-topic send should be retried with routing stripped.
-
-        Two cases trigger the retry:
-
-        1. The original anchor-stale case — the reply target was deleted, so
-           Bot API returns "message to be replied not found". The retry drops
-           the reply anchor and the topic id together.
-
-        2. The synthetic-event case (added when #27937 introduced
-           ``direct_messages_topic_id`` fallback for sends without an anchor):
-           if Bot API rejects the topic id itself with any BadRequest that
-           mentions topic/thread routing, we retry without routing rather
-           than dropping the message.
-        """
-        if not (metadata and metadata.get("telegram_dm_topic_reply_fallback")):
-            return False
-        if not cls._is_bad_request_error(error):
-            return False
-        err_lower = str(error).lower()
-        if reply_to_message_id is not None and "message to be replied not found" in err_lower:
-            return True
-        # Synthetic / resumed sends route via ``direct_messages_topic_id``
-        # instead of a reply anchor. If Telegram rejects the topic id, fall
-        # back to a plain DM send.
-        if metadata.get("direct_messages_topic_id"):
-            topic_markers = (
-                "direct_messages_topic",
-                "message thread not found",
-                "thread not found",
-                "topic_closed",
-                "topic_deleted",
-                "topic not found",
-            )
-            if any(marker in err_lower for marker in topic_markers):
-                return True
-        return False
-
-    async def _send_with_dm_topic_reply_anchor_retry(
-        self,
-        send_fn: Any,
-        send_kwargs: Dict[str, Any],
-        metadata: Optional[Dict[str, Any]],
-        reply_to_message_id: Optional[int],
-        media_label: str,
-        reset_media: Optional[Any] = None,
-    ) -> Any:
-        """Retry stale private-topic media replies once without the topic anchor."""
-        try:
-            return await send_fn(**send_kwargs)
-        except Exception as send_err:
-            if not self._should_retry_without_dm_topic_reply_anchor(
-                send_err,
-                metadata,
-                reply_to_message_id,
-            ):
-                raise
-            logger.warning(
-                "[%s] Reply target deleted for Telegram %s, "
-                "retrying without reply/topic anchor: %s",
-                self.name,
-                media_label,
-                _redact_telegram_error_text(send_err),
-            )
-            if reset_media is not None:
-                reset_media()
-            retry_kwargs = dict(send_kwargs)
-            retry_kwargs["reply_to_message_id"] = None
-            retry_kwargs.pop("message_thread_id", None)
-            retry_kwargs.pop("direct_messages_topic_id", None)
-            return await send_fn(**retry_kwargs)
-
-    def _fallback_ips(self) -> list[str]:
-        """Return validated fallback IPs from config (populated by _apply_env_overrides)."""
-        configured = self.config.extra.get("fallback_ips", []) if getattr(self.config, "extra", None) else []
-        if isinstance(configured, str):
-            configured = configured.split(",")
-        return parse_fallback_ip_env(",".join(str(v) for v in configured) if configured else None)
-
-    @staticmethod
-    def _looks_like_polling_conflict(error: Exception) -> bool:
-        text = str(error).lower()
-        return (
-            error.__class__.__name__.lower() == "conflict"
-            or "terminated by other getupdates request" in text
-            or "another bot instance is running" in text
-        )
-
-    @staticmethod
-    def _looks_like_network_error(error: Exception) -> bool:
-        """Return True for transient transport failures that warrant reconnect."""
-        name = error.__class__.__name__.lower()
-        if name in {"badrequest", "invalidtoken", "forbidden", "retryafter"}:
-            return False
-        if name in {"networkerror", "timedout", "connectionerror"}:
-            return True
-        try:
-            from telegram.error import (
-                BadRequest,
-                Forbidden,
-                InvalidToken,
-                NetworkError,
-                RetryAfter,
-                TimedOut,
-            )
-            if isinstance(error, (BadRequest, InvalidToken, Forbidden, RetryAfter)):
-                return False
-            if isinstance(error, (NetworkError, TimedOut)):
-                return True
-        except ImportError:
-            pass
-        return isinstance(error, OSError)
-
-    @staticmethod
-    def _looks_like_connect_timeout(error: Exception) -> bool:
-        """Return True when a Telegram TimedOut wraps a connect-timeout.
-
-        A plain Telegram TimedOut may mean the request reached Telegram and
-        should not be re-sent. A ConnectTimeout means the TCP connection was
-        never established, so retrying is safe and prevents silent drops.
-        """
-        seen: set[int] = set()
-        stack: list[BaseException] = [error]
-        while stack:
-            cur = stack.pop()
-            ident = id(cur)
-            if ident in seen:
-                continue
-            seen.add(ident)
-            name = cur.__class__.__name__.lower()
-            text = str(cur).lower()
-            if "connecttimeout" in name or "connect timeout" in text or "connect timed out" in text:
-                return True
-            cause = getattr(cur, "__cause__", None)
-            context = getattr(cur, "__context__", None)
-            if cause is not None:
-                stack.append(cause)
-            if context is not None:
-                stack.append(context)
-        return False
-
-    @staticmethod
-    def _looks_like_pool_timeout(error: Exception) -> bool:
-        """Return True when a Telegram TimedOut wraps an httpx pool timeout.
-
-        PTB converts ``httpx.PoolTimeout`` into ``telegram.error.TimedOut`` with
-        a message that explicitly states the request was *not* sent
-        (``"Pool timeout: All connections in the connection pool are occupied.
-        Request was *not* sent to Telegram."``). Because the request never left
-        the process, re-sending is safe and cannot duplicate -- the opposite of
-        a generic TimedOut, which may have reached Telegram. We match the
-        wrapped ``httpx.PoolTimeout`` class as well as the message string so the
-        check survives PTB message-wording changes.
-        """
-        seen: set[int] = set()
-        stack: list[BaseException] = [error]
-        while stack:
-            cur = stack.pop()
-            ident = id(cur)
-            if ident in seen:
-                continue
-            seen.add(ident)
-            name = cur.__class__.__name__.lower()
-            text = str(cur).lower()
-            if "pooltimeout" in name or "pool timeout" in text or (
-                "connection pool" in text and "occupied" in text
-            ):
-                return True
-            cause = getattr(cur, "__cause__", None)
-            context = getattr(cur, "__context__", None)
-            if cause is not None:
-                stack.append(cause)
-            if context is not None:
-                stack.append(context)
-        return False
-
-    def _coerce_bool_extra(self, key: str, default: bool = False) -> bool:
-        value = self.config.extra.get(key) if getattr(self.config, "extra", None) else None
-        if value is None:
-            return default
-        if isinstance(value, str):
-            lowered = value.strip().lower()
-            if lowered in {"true", "1", "yes", "on"}:
-                return True
-            if lowered in {"false", "0", "no", "off"}:
-                return False
-            return default
-        return bool(value)
-
-    def _coerce_float_extra(
-        self,
-        key: str,
-        default: float,
-        *,
-        min_value: Optional[float] = None,
-        max_value: Optional[float] = None,
-    ) -> float:
-        value = self.config.extra.get(key) if getattr(self.config, "extra", None) else None
-        if value is None:
-            return default
-        try:
-            parsed = float(value)
-        except (TypeError, ValueError):
-            return default
-        if min_value is not None:
-            parsed = max(parsed, min_value)
-        if max_value is not None:
-            parsed = min(parsed, max_value)
-        return parsed
-
     def _link_preview_kwargs(self) -> Dict[str, Any]:
         if not getattr(self, "_disable_link_previews", False):
             return {}
         if LinkPreviewOptions is not None:
             return {"link_preview_options": LinkPreviewOptions(is_disabled=True)}
         return {"disable_web_page_preview": True}
-
-    # ------------------------------------------------------------------
-    # Bot API 10.1 Rich Messages (sendRichMessage)
-    #
-    # Final / new-message replies opportunistically use sendRichMessage with
-    # the RAW agent markdown so richer constructs (tables, task lists,
-    # collapsible details, math, ...) render natively. The legacy MarkdownV2
-    # send() path stays as the fallback for unsupported/oversized content and
-    # older PTB/clients. Streaming edits stay on Hermes' existing MarkdownV2
-    # edit path for now; finalization can re-send as rich and delete the stale
-    # preview until rich_message edit support is wired directly.
-    # ------------------------------------------------------------------
-    def _content_fits_rich_limits(self, content: str) -> bool:
-        """Cheap pre-check for the one hard rich limit we can count locally.
-
-        Only the 32,768 UTF-8 character text cap is enforced here. Other Bot API
-        rich limits (500 blocks, 16 nesting levels, 20 table columns, ...) are
-        not pre-counted; if exceeded Telegram returns a BadRequest, which
-        :meth:`_is_rich_fallback_error` classifies as permanent so the send
-        degrades to the legacy chunking path.
-        """
-        return len(content) <= self.RICH_MESSAGE_MAX_CHARS
-
-    def _bot_supports_rich(self) -> bool:
-        """True when the bound bot can issue raw ``sendRichMessage`` calls.
-
-        Gates on ``do_api_request`` being an *async* callable. The real
-        ``telegram.Bot.do_api_request`` is a coroutine function; test doubles
-        that opt into rich set it to an ``AsyncMock`` (also a coroutine
-        function). Plain ``MagicMock`` bots expose a *sync* auto-child and
-        ``SimpleNamespace`` bots lack the attribute entirely — both resolve to
-        ``False`` here, so the legacy path is used unchanged.
-        """
-        return inspect.iscoroutinefunction(getattr(self._bot, "do_api_request", None))
-
-    _RICH_DETAILS_RE = re.compile(r"<details\b[^>]*>.*?</details>", re.IGNORECASE | re.DOTALL)
-    _RICH_MATH_IN_DETAILS_RE = re.compile(
-        r"(\$\$.*?\$\$|"
-        r"\\\[.*?\\\]|"
-        r"\\\(.*?\\\)|"
-        r"\\(?:sum|frac|alpha|beta|gamma|delta|theta|lambda|mu|pi|sigma|"
-        r"int|prod|sqrt|lim|infty|begin\{(?:equation|align|matrix|cases)\}))",
-        re.IGNORECASE | re.DOTALL,
-    )
-    _RICH_CJK_RE = re.compile(
-        "["
-        "\u3040-\u30ff"  # Hiragana, Katakana
-        "\u3400-\u4dbf"  # CJK Extension A
-        "\u4e00-\u9fff"  # CJK Unified Ideographs
-        "\uac00-\ud7af"  # Hangul syllables
-        "\uf900-\ufaff"  # CJK Compatibility Ideographs
-        "\U00020000-\U000323af"  # CJK extensions and compatibility supplement
-        "]"
-    )
-
-    def _has_telegram_desktop_details_math_crash_shape(self, content: str) -> bool:
-        """Return True for rich-message details+math content that crashes TDesktop.
-
-        Telegram Desktop 6.9.1 can crash while rendering Bot API 10.1 rich
-        messages containing math inside a collapsible details block
-        (telegramdesktop/tdesktop#30808). The Bot API accepts the payload, so
-        Hermes must skip rich delivery up front and use the legacy MarkdownV2
-        path until affected Desktop clients age out.
-        """
-        if not content:
-            return False
-        for details_block in self._RICH_DETAILS_RE.findall(content):
-            if self._RICH_MATH_IN_DETAILS_RE.search(details_block):
-                return True
-        return False
-
-    def _has_telegram_desktop_cjk_rich_garble_shape(self, content: str) -> bool:
-        """Return True for CJK content that current TDesktop rich drafts garble.
-
-        Telegram Mac/Desktop Bot API 10.1 rich-message rendering currently
-        leaves overlapping draft/overlay glyph artifacts for CJK text (#47653).
-        The legacy MarkdownV2 path renders the same text cleanly, so skip rich
-        delivery up front until affected clients age out.
-        """
-        return bool(content and self._RICH_CJK_RE.search(content))
-
-    def _needs_rich_rendering(self, content: str) -> bool:
-        """Return True for markdown constructs that the legacy path degrades.
-
-        Keep ordinary replies on the pre-rich MarkdownV2 path so Telegram
-        clients render a consistent font weight/spacing. The rich endpoint is
-        reserved for constructs where raw markdown materially improves output:
-        pipe tables (MarkdownV2 has no table syntax and rewrites them into
-        bullet lists), GFM task lists, collapsible ``<details>`` blocks, and
-        block math.  Adapted from #45995 (@YonganZhang).
-        """
-        if not content:
-            return False
-        if any(_TABLE_SEPARATOR_RE.match(line) for line in content.splitlines()):
-            return True
-        if re.search(r"(?m)^\s*[-*]\s+\[[ xX]\]\s+", content):
-            return True
-        if re.search(r"(?m)^<details\b|^</details>|^<summary\b|^</summary>", content):
-            return True
-        if "$$" in content:
-            return True
-        return False
-
-    def _rich_delivery_enabled(self) -> bool:
-        """Whether rich delivery is allowed (``rich_messages`` opt-in)."""
-        return bool(getattr(self, "_rich_messages_enabled", True))
-
-    def _rich_eligible(self, content: str) -> bool:
-        """Capability/content eligibility for rich, ignoring ``expect_edits``.
-
-        Shared core of :meth:`_should_attempt_rich` minus the per-call
-        ``expect_edits`` metadata gate.  The rich EDIT-finalize path
-        (:meth:`_try_edit_rich`) needs this: a streamed preview is sent with
-        ``expect_edits=True`` to stay on the editable path mid-stream, but the
-        FINAL edit should still upgrade to rich when the content warrants it.
-        """
-        return bool(
-            self._rich_delivery_enabled()
-            and not getattr(self, "_rich_send_disabled", False)
-            and content
-            and content.strip()
-            and self._needs_rich_rendering(content)
-            and not self._has_telegram_desktop_details_math_crash_shape(content)
-            and not self._has_telegram_desktop_cjk_rich_garble_shape(content)
-            and self._content_fits_rich_limits(content)
-            and self._bot_supports_rich()
-        )
-
-    def _should_attempt_rich(
-        self, content: str, metadata: Optional[Dict[str, Any]] = None
-    ) -> bool:
-        return bool(
-            not (metadata or {}).get("expect_edits")
-            and self._rich_eligible(content)
-        )
-
-    def prefers_fresh_final_streaming(
-        self, content: str, metadata: Optional[Dict[str, Any]] = None
-    ) -> bool:
-        """Whether to replace a streamed preview with a fresh rich final.
-
-        Disabled for Telegram. The fresh-final path briefly shows two copies of
-        the final answer, then deletes the streaming preview after the rich send
-        succeeds — it looks like duplicate delivery at the end of every streamed
-        turn (the reason #46206 reverted it).  Rich finalize is instead handled
-        by editing the existing preview in place via Bot API 10.1's
-        ``editMessageText`` ``rich_message`` parameter (see
-        :meth:`_try_edit_rich`), so no fresh re-send / delete is needed.
-        """
-        return False
-
-    def streaming_overflow_limit(self) -> Optional[int]:
-        """Allow the stream consumer to accumulate up to the rich-message cap
-        before splitting, so a reply that fits one ``sendRichMessage`` /
-        ``sendRichMessageDraft`` isn't fragmented at the 4,096 MarkdownV2 limit.
-
-        Gated on the same rich capability as the send path (minus the
-        content-length check — raising that cap is the whole point): rich not
-        latched off and the bot exposes an async ``do_api_request``.  Returns
-        ``None`` (→ legacy 4,096 limit) when rich isn't available, so non-rich
-        streams split exactly as before.
-        """
-        if (
-            getattr(self, "_rich_messages_enabled", True)
-            and not getattr(self, "_rich_send_disabled", False)
-            and self._bot_supports_rich()
-        ):
-            return self.RICH_MESSAGE_MAX_CHARS
-        return None
-
-    def _rich_message_payload(
-        self, content: str, *, skip_entity_detection: bool = False
-    ) -> Dict[str, Any]:
-        """Build the ``InputRichMessage`` object from RAW markdown.
-
-        Never pass ``format_message(content)`` here — that converts to
-        MarkdownV2 and would escape/destroy rich syntax like table pipes.
-
-        Single newlines are normalized to Markdown hard breaks so that
-        multi-line content (slash-command lists, etc.) renders correctly
-        in the rich-message path.  See ``_rich_normalize_linebreaks``.
-        """
-        payload: Dict[str, Any] = {"markdown": _rich_normalize_linebreaks(content)}
-        if skip_entity_detection:
-            payload["skip_entity_detection"] = True
-        return payload
-
-    def _is_rich_capability_error(self, exc: Exception) -> bool:
-        """True ⇒ the rich endpoint itself is unavailable (old PTB/server).
-
-        These latch rich off for the rest of the adapter's life — retrying is
-        pointless and would cost a failed roundtrip on every send. Per-message
-        rejections (BadRequest from a parser/limit issue) are NOT capability
-        errors: the next message may be fine.
-        """
-        name = exc.__class__.__name__.lower()
-        if name in {"endpointnotfound", "invalidtoken"}:
-            return True
-        if isinstance(exc, (AttributeError, TypeError, NotImplementedError)):
-            return True
-        if getattr(exc, "error_code", None) == 404:
-            return True
-        s = str(exc).lower()
-        if ("method" in s or "endpoint" in s) and (
-            "not found" in s or "does not exist" in s
-        ):
-            return True
-        return "no such method" in s
-
-    def _is_rich_fallback_error(self, exc: Exception) -> bool:
-        """True ⇒ permanent/capability error ⇒ safe to fall back to legacy.
-
-        Conservative on purpose: only clearly-permanent failures (BadRequest,
-        capability errors, unknown/unsupported endpoint) qualify. Everything
-        else is treated as transient — the rich request may have reached
-        Telegram, so we must NOT legacy-resend and risk a duplicate.
-        """
-        if self._is_bad_request_error(exc):
-            return True
-        if self._is_rich_capability_error(exc):
-            return True
-        s = str(exc).lower()
-        return "unsupported" in s or "not implemented" in s
-
-    def _compute_single_send_routing(
-        self,
-        chat_id: str,
-        reply_to: Optional[str],
-        metadata: Optional[Dict[str, Any]],
-        thread_id: Optional[str],
-    ) -> Optional[tuple]:
-        """Routing for a single (rich) send — mirrors send()'s index-0 block.
-
-        Returns ``(reply_to_id, thread_kwargs)``, or ``None`` to signal "skip
-        rich, let the legacy path handle it" — used for the DM-topic fail-loud
-        case so the legacy path stays the single source of the refuse result.
-        """
-        metadata_reply_to = self._metadata_reply_to_message_id(metadata)
-        private_dm_topic_send = self._is_private_dm_topic_send(chat_id, thread_id, metadata)
-        dm_topic_reply_to_off = (
-            private_dm_topic_send
-            and self._reply_to_mode == "off"
-            and bool(metadata and metadata.get("telegram_dm_topic_reply_fallback"))
-        )
-        reply_to_source = reply_to or (
-            str(metadata_reply_to)
-            if private_dm_topic_send and metadata_reply_to is not None
-            else None
-        )
-        if private_dm_topic_send:
-            should_thread = reply_to_source is not None and self._reply_to_mode != "off"
-        else:
-            should_thread = self._should_thread_reply(reply_to_source, 0)
-        reply_to_id = int(reply_to_source) if should_thread and reply_to_source else None
-        thread_kwargs = self._thread_kwargs_for_send(
-            chat_id,
-            thread_id,
-            metadata,
-            reply_to_message_id=reply_to_id,
-            reply_to_mode=self._reply_to_mode,
-        )
-        if private_dm_topic_send and reply_to_id is None and not dm_topic_reply_to_off:
-            # Refusing to send outside the requested DM topic — defer to the
-            # legacy path, which returns the canonical fail-loud SendResult.
-            # Exception: synthetic/resumed topic sends that route via
-            # ``direct_messages_topic_id`` do not need a reply anchor.
-            if not thread_kwargs.get("direct_messages_topic_id"):
-                return None
-        return reply_to_id, thread_kwargs
-
-    async def _try_send_rich(
-        self,
-        chat_id: str,
-        content: str,
-        reply_to: Optional[str],
-        metadata: Optional[Dict[str, Any]],
-    ) -> Optional[SendResult]:
-        """Attempt a single ``sendRichMessage`` send.
-
-        Returns a :class:`SendResult` (success, or a transient failure that the
-        caller must NOT legacy-resend), or ``None`` to signal "fall back to the
-        legacy MarkdownV2 path" (permanent/capability error or DM-topic skip).
-        """
-        thread_id = self._metadata_thread_id(metadata)
-        routing = self._compute_single_send_routing(chat_id, reply_to, metadata, thread_id)
-        if routing is None:
-            return None
-        reply_to_id, thread_kwargs = routing
-
-        payload: Dict[str, Any] = {
-            "chat_id": normalize_telegram_chat_id(chat_id),
-            "rich_message": self._rich_message_payload(content),
-        }
-        # Only forward non-None routing keys: when direct_messages_topic_id is
-        # present _thread_kwargs_for_send pairs it with message_thread_id=None,
-        # which must not be sent as a stray field on the raw endpoint.
-        payload.update({k: v for k, v in thread_kwargs.items() if v is not None})
-        payload.update(self._notification_kwargs(metadata))
-        if getattr(self, "_disable_link_previews", False):
-            payload["link_preview_options"] = {"is_disabled": True}
-        if reply_to_id is not None:
-            # Spec: sendRichMessage takes reply_parameters (ReplyParameters
-            # object), NOT the legacy reply_to_message_id scalar. Unknown
-            # params are silently ignored by the Bot API, so the scalar would
-            # quietly drop the reply anchor instead of erroring.
-            payload["reply_parameters"] = {"message_id": reply_to_id}
-
-        try:
-            # Take the raw Bot API result (dict under real PTB). Passing
-            # return_type=Message would make PTB deserialize a Bot API 10.1
-            # response shape it does not fully model yet; a post-delivery parse
-            # error must not be mistaken for a sendable failure.
-            msg = await self._bot.do_api_request(
-                "sendRichMessage", api_kwargs=payload
-            )
-        except Exception as exc:
-            if self._is_rich_fallback_error(exc):
-                if self._is_rich_capability_error(exc):
-                    # Endpoint missing (old PTB/server) — latch rich off so
-                    # every later send doesn't pay a doomed extra roundtrip.
-                    self._rich_send_disabled = True
-                logger.debug(
-                    "[%s] sendRichMessage rejected (%s) — falling back to MarkdownV2",
-                    self.name, _redact_telegram_error_text(exc),
-                )
-                return None
-            # Transient / network / unknown: the request may have reached
-            # Telegram. Do NOT legacy-resend (duplicate risk); surface a
-            # failure with retry semantics mirroring the legacy send() except.
-            err_str = str(exc).lower()
-            try:
-                from telegram.error import TimedOut as _TimedOut
-            except (ImportError, AttributeError):
-                _TimedOut = None
-            is_timeout = (_TimedOut and isinstance(exc, _TimedOut)) or "timed out" in err_str
-            is_connect_timeout = self._looks_like_connect_timeout(exc)
-            # Extract server-requested retry_after for flood control so the
-            # base retry layer honors Telegram's backoff instead of its own
-            # short exponential schedule.
-            _retry_after = getattr(exc, "retry_after", None)
-            if _retry_after is None:
-                import re as _re
-                _m = _re.search(r"retry\s+(?:in\s+)?(\d+)", err_str, _re.IGNORECASE)
-                if _m:
-                    _retry_after = float(_m.group(1))
-            safe_error = _redact_telegram_error_text(exc)
-            logger.warning(
-                "[%s] sendRichMessage transient failure (no legacy resend): %s",
-                self.name, safe_error,
-            )
-            return SendResult(
-                success=False,
-                error=safe_error,
-                retryable=(is_connect_timeout or not is_timeout),
-                retry_after=_retry_after,
-            )
-
-        message_id = None
-        if isinstance(msg, dict):
-            message_id = msg.get("message_id")
-            if message_id is None:
-                message_id = (msg.get("result") or {}).get("message_id")
-        else:
-            message_id = getattr(msg, "message_id", None)
-        if message_id is not None:
-            # Telegram won't echo rich content in reply_to_message, so remember
-            # what we sent — replies to this message resolve via this index.
-            try:
-                from gateway import rich_sent_store
-                rich_sent_store.record(str(chat_id), str(message_id), content)
-            except Exception:
-                pass
-        return SendResult(
-            success=True,
-            message_id=str(message_id) if message_id is not None else None,
-        )
-
-    async def _try_edit_rich(
-        self,
-        chat_id: str,
-        message_id: str,
-        content: str,
-        metadata: Optional[Dict[str, Any]] = None,
-    ) -> Optional[SendResult]:
-        """Edit an existing message in place as a rich message (Bot API 10.1).
-
-        Uses ``editMessageText`` with the ``rich_message`` parameter so a
-        streamed preview can finalize as rich (tables/task lists/details/math)
-        WITHOUT a fresh send + delete — no duplicate preview.  Mirrors
-        :meth:`_try_send_rich`'s error contract:
-
-        - success → ``SendResult(success=True, message_id=...)``
-        - permanent / capability error → ``None`` (caller falls back to the
-          legacy MarkdownV2 edit; capability errors latch rich off)
-        - transient / unknown → ``SendResult(success=False)`` with retry
-          semantics (the message may already be edited; do NOT legacy-resend)
-        """
-        payload: Dict[str, Any] = {
-            "chat_id": normalize_telegram_chat_id(chat_id),
-            "message_id": int(message_id),
-            "rich_message": self._rich_message_payload(content),
-        }
-        thread_id = self._metadata_thread_id(metadata)
-        thread_kwargs = self._thread_kwargs_for_send(
-            chat_id,
-            thread_id,
-            metadata,
-            reply_to_message_id=None,
-            reply_to_mode=self._reply_to_mode,
-        )
-        payload.update({k: v for k, v in thread_kwargs.items() if v is not None})
-        if getattr(self, "_disable_link_previews", False):
-            payload["link_preview_options"] = {"is_disabled": True}
-        try:
-            # Raw Bot API result; do not request return_type=Message (PTB does
-            # not fully model the 10.1 response shape yet — a post-edit parse
-            # error must not be mistaken for a failed edit).
-            await self._bot.do_api_request("editMessageText", api_kwargs=payload)
-        except Exception as exc:
-            if self._is_rich_fallback_error(exc):
-                if self._is_rich_capability_error(exc):
-                    self._rich_send_disabled = True
-                # "Message is not modified" — content identical to the current
-                # rich message; treat as a successful no-op so the caller does
-                # not fall through to a redundant legacy edit.
-                if "not modified" in str(exc).lower():
-                    return SendResult(success=True, message_id=message_id)
-                logger.debug(
-                    "[%s] rich editMessageText rejected (%s) — falling back to MarkdownV2 edit",
-                    self.name, _redact_telegram_error_text(exc),
-                )
-                return None
-            if "not modified" in str(exc).lower():
-                return SendResult(success=True, message_id=message_id)
-            err_str = str(exc).lower()
-            try:
-                from telegram.error import TimedOut as _TimedOut
-            except (ImportError, AttributeError):
-                _TimedOut = None
-            is_timeout = (_TimedOut and isinstance(exc, _TimedOut)) or "timed out" in err_str
-            is_connect_timeout = self._looks_like_connect_timeout(exc)
-            safe_error = _redact_telegram_error_text(exc)
-            logger.warning(
-                "[%s] rich editMessageText transient failure (no legacy resend): %s",
-                self.name, safe_error,
-            )
-            return SendResult(
-                success=False,
-                error=safe_error,
-                retryable=(is_connect_timeout or not is_timeout),
-            )
-        # Telegram won't echo rich content for messages that predate the bot's
-        # first rich send, so mirror the fresh-send index here too: a streamed
-        # final finalized via editMessageText is otherwise never recorded, and
-        # replies to it would have no native echo to recover from.
-        try:
-            from gateway import rich_sent_store
-            rich_sent_store.record(str(chat_id), str(message_id), content)
-        except Exception:
-            pass
-        return SendResult(success=True, message_id=message_id)
 
     def _should_attempt_rich_draft(self, content: str) -> bool:
         return bool(
